+++ pyppl/__init__.py
@@ -1,6 +1,6 @@
 """PyPPL - A Python PiPeLine framework."""
 
-__version__ = "3.0.3"
+__version__ = "3.1.0"
 
 from .config import load_config
 from .plugin import config_plugins